% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cleaninglog.R
\name{cleaninglog}
\alias{cleaninglog}
\title{create standardised cleaninglog}
\usage{
cleaninglog(ids, variables, new_values, data_id_column_name,
  name = "default_cleaninglog_element_name", info = NULL, ...)
}
\arguments{
\item{ids}{vector of ids that can be matched with the dataset (i.e. uuids)}

\item{new_values}{vector of new (corrected) values. If no change should be made, this must be set to the original value}

\item{data_id_column_name}{the name of the column in the dataset with ids matching the values in `ids`}

\item{name}{a vector name of the data check that produced the}

\item{info}{(optional): a character vector with additional information about the cleaning log}

\item{...}{additional vectors or data.frames to be added to the cleaning log.}

\item{variabels}{vector of variable names (should match data column headers)}
}
\value{
a `clog` cleaninglog objects to be used with `clog_clean()` etc.
}
\description{
create standardised cleaninglog
}
