% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cleaninglog.R
\name{clog_clean}
\alias{clog_clean}
\title{make changes to raw dataset based on cleaninglog (see `cleaninglog()` )}
\usage{
clog_clean(df, cleaninglog)
}
\arguments{
\item{df}{a data.frame / tibble}

\item{cleaninglog}{a cleaninglog object create with `cleaninglog()`}
}
\value{
the original dataframe (`df`) modified based on the cleaning log.
}
\description{
make changes to raw dataset based on cleaninglog (see `cleaninglog()` )
}
