% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaninglog_data_exchange.R
\name{clog_cleaninglog_to_data}
\alias{clog_cleaninglog_to_data}
\title{add cleaninglog elements as columns to a datset}
\usage{
clog_cleaninglog_to_data(data, cleaninglog)
}
\arguments{
\item{data}{the dataset to add cleaninglog to as columns}

\item{cl}{cleaninglog}
}
\value{
the original `data` with one additional column per unique value cleaninglog `name`. The `name` becomes the column name. For each row in the data, the new column is `TRUE` if a cleaninglog element with that name was listed for the respective row
}
\description{
add cleaninglog elements as columns to a datset
}
