% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaninglog_data_exchange.R
\name{clog_data_to_cleaninglog}
\alias{clog_data_to_cleaninglog}
\title{add cleaninglog elements as columns to a datset}
\usage{
clog_data_to_cleaninglog(data, cleaninglog, ...)
}
\arguments{
\item{data}{the dataset to select variabes from that should be added to the cleaning log}

\item{cleaninglog}{the cleaninglog (see `cleaninglog()` )}

\item{...}{names of the variables to be added from the dataset to the cleaninglog (without quotes!)}
}
\value{
the original `data` with one additional column per unique value cleaninglog `name`. The `name` becomes the column name. For each row in the data, the new column is `TRUE` if a cleaninglog element with that name was listed for the respective row
}
\description{
add cleaninglog elements as columns to a datset
}
\details{
returning `NA`` if the id in the cleaning log can not be found in the data
}
