% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cleaninglog.R
\name{clog_get_cleaninglog}
\alias{clog_get_cleaninglog}
\title{extract cleaninglog from dataset modified with `clog_clean()`}
\usage{
clog_get_cleaninglog(df)
}
\arguments{
\item{df}{a dataset modified with `clog_clean()`}
}
\value{
the cleaninglog as it was applied to `df`
}
\description{
extract cleaninglog from dataset modified with `clog_clean()`
}
