% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cleaninglog.R
\name{clog_reverse}
\alias{clog_reverse}
\title{reverse changes applied based on cleaning log}
\usage{
clog_reverse(df)
}
\arguments{
\item{df}{a data frame modified with `clog_clean()`}
}
\value{
the original cleaning log applied in reverse.
}
\description{
reverse changes applied based on cleaning log
}
